- Fix #1; - Rename the Listener and improve it with onTick(); - Refactor and clean all lib code;
- Add 'kotlin-android' to fix problems;
- Rename 'startTime' to 'pastTime' for better understanding;
- Separete the 'listener' from 'create';
- 'start()' now return the instance;

+++ app/src/main/java/douglasspgyn/com/github/circularcountdown/ProgressBarAnimation.kt
@@ -12,6 +12,6 @@ class ProgressBarAnimation(private val progressBar: CircularCountdown, private v
     override fun applyTransformation(interpolatedTime: Float, t: Transformation) {
         super.applyTransformation(interpolatedTime, t)
         val value = from + (to - from) * interpolatedTime
-        progressBar.setProgress(value.toInt())
+        progressBar.setProgress(value.toLong())
     }
 }
